requirements.txt
# Agentic AI Data Harmonization System
# Requirements File

# Core Dependencies
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Format Support
openpyxl>=3.1.0
pyreadstat>=1.2.0
xlrd>=2.0.0

# Reporting
jinja2>=3.1.0
plotly>=5.15.0

# Vector Memory (Stretch Goal)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Streamlit UI (Stretch Goal)
streamlit>=1.28.0

# Logging and Utilities
rich>=13.0.0
tabulate>=0.9.0

# Testing
pytest>=7.4.0

